changes in this commit: - some bug fixes

+++ database/migrations/2025_02_28_102626_create_main_product_tables.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('normalized_name');
             $table->string('code')->nullable();
-            $table->decimal('quantity', 15, 2)->nullable();
+            $table->integer('quantity')->nullable();
             $table->decimal('price', 15, 2)->nullable();
             $table->text('description')->nullable();
             $table->string('sheet_name');
